+++ tensorrt_llm/_torch/auto_deploy/custom_ops/__init__.py
@@ -8,7 +8,7 @@ from .fused_moe import *
 from .linear import *
 from .mla import *
 from .quant import *
-from .rope import *
+from .triton_rope import *
 from .torch_attention import *
 from .torch_rope import *
 from .triton_attention import *
